+++ app/components/exploreFurther/exploreFurther.tsx
@@ -9,13 +9,13 @@ import {
   CardContent,
   Tooltip,
 } from "@nypl/design-system-react-components";
-import { ExploreFurtherData } from "../../types/ExploreFurtherData";
+import { ExploreFurtherDataType } from "../../types/ExploreFurtherData";
 import exploreFurtherData from "../../data/exploreFurtherData";
 import { headerBreakpoints } from "../../utils/breakpoints";
 import useBreakpoints from "../../hooks/useBreakpoints";
 
 const ExploreFurther = () => {
-  const data: ExploreFurtherData[] = exploreFurtherData;
+  const data: ExploreFurtherDataType[] = exploreFurtherData;
   const { isLargerThanLargeTablet } = useBreakpoints();
   function exploreFurtherLink(item) {
     return (