+++ app/components/swimlanes/swimLanes.tsx
@@ -14,9 +14,11 @@ import {
   Tooltip,
 } from "@nypl/design-system-react-components";
 import styles from "./Swimlanes.module.css";
-import { imageURL } from "../../utils/utils";
 import { DC_URL } from "../../config/constants";
-import useBreakpoints from "app/hooks/useBreakpoints";
+import useBreakpoints from "../../hooks/useBreakpoints";
+import CollectionCard from "../cards/collectionCard";
+import type { CollectionCardDataType } from "../../types/CollectionCard";
+import { CollectionCardModel } from "../../models/collectionCard";
 
 const SwimLanes = ({ numColumns, lanesWithNumItems }) => {
   const { isLargerThanLargeTablet } = useBreakpoints();
@@ -53,67 +55,18 @@ const SwimLanes = ({ numColumns, lanesWithNumItems }) => {
             gridTemplateColumns: `repeat(${numColumns}, minmax(0, 1fr))`,
           }}
         >
-          {lane.collections.map((collection, index) => (
-            <Card
-              sx={{ display: "grid" }}
-              key={index}
-              id={`card-${lane.slug}-${index}`}
-              mainActionLink={collection.url}
-              imageProps={{
-                alt: "",
-                id: `image-${lane.slug}-${index}`,
-                isLazy: true,
-                aspectRatio: "twoByOne",
-                src: imageURL(collection.image_id, "full", "288,", "0"),
-                // TODO: *IF* we want to use the Nextjs Image component, this
-                // is how we would do it. It's suppose to be better for
-                // performance but it's not visibly noticeable.
-                // component: (
-                //   <Image
-                //     src={imageURL(collection.image_id, "full", "288,", "0")}
-                //     alt=""
-                //     width={100}
-                //     height={100}
-                //     objectFit="cover"
-                //   />
-                // ),
-              }}
-            >
-              <CardHeading
-                id={`row-card-heading-${lane.slug}-${index}`}
-                level="h3"
-                size="heading5"
-                className={styles.collectiontitle}
-                noOfLines={3}
-              >
-                {isLargerThanLargeTablet ? (
-                  <Tooltip content={collection.title}>
-                    <Link href={collection.url} sx={{ marginBottom: "0" }}>
-                      {collection.title}
-                    </Link>
-                  </Tooltip>
-                ) : (
-                  collection.title
-                )}
-              </CardHeading>
-              <CardContent>
-                <Text
-                  id={`item-count-${lane.slug}-${index}`}
-                  size="subtitle2"
-                  fontWeight="medium"
-                  __css={{
-                    display: "none",
-                    [`@media screen and (min-width: 480px)`]: {
-                      display: "inline",
-                    },
-                  }}
-                >
-                  {" "}
-                  {`${collection.numItems || 0} items`}{" "}
-                </Text>
-              </CardContent>
-            </Card>
-          ))}
+          {lane.collections.map((collection: CollectionCardDataType, index) => {
+            const c = new CollectionCardModel(collection); // can remove this and the line above after referencing the CollectionCardModel in the Lane model after it's created
+            return (
+              <CollectionCard
+                key={index}
+                slug={lane.slug}
+                id={index}
+                collection={c} //rename to collection after Lane model is created
+                isLargerThanLargeTablet={isLargerThanLargeTablet}
+              />
+            );
+          })}
         </SimpleGrid>
       )}
       <Link