+++ app/utils/utils.test.tsx
@@ -126,7 +126,7 @@ describe("getItemsCountFromUUIDs()", () => {
   });
 });
 
-// TO DO:
+// TODO:
 /**
  * Represents a IIIF Image API URL, which will be used globally throughout the application.
  * IIIF Image API has several params, the ones we are the most concerned about are Region, Size, and Rotation.