+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Added `/search/index` and `/collections/slug` pages (DR-3028)
+
 ## [0.1.9] 2024-07-08
 
 ### Updated