+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updates card columns to new breakpoints (DR-3036)
 - Migrated to Next v14 (App Router) (DR-3010, DR-3011)
 - Make the entire card clickable for Collection and Explore Further cards (DR-3024)
+- Create CollectionCard component and render it on the homepage, also refactor a little bit. (DR-3027)
 
 ### Added