+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `/items` pages (DR-3029)
 - Added `/api/changelog` endpoint (DR-3048)
 
+### Updated
+
+- Updated static breadcrumbs to render correct copy (DR-3023)
+
 ## [0.1.9] 2024-07-08
 
 ### Updated