+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `/search/index` and `/collections/slug` pages (DR-3028)
 - Added `/items` pages (DR-3029)
+- Added Github Actions deployment workflows (DR-3019)
 
 ## [0.1.9] 2024-07-08