+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.1.10] 2024-07-08
+
 ### Added
 
 - Added `/search/index` and `/collections/slug` pages (DR-3028)